+++ app/angular/src/builders/build-storybook/index.spec.ts
@@ -84,6 +84,7 @@ describe('Build Storybook Builder', () => {
       outputDir: 'storybook-static',
       mode: 'static',
       tsConfig: './storybook/tsconfig.ts',
+      webpackStatsJson: false,
     });
   });
 
@@ -109,11 +110,39 @@ describe('Build Storybook Builder', () => {
       outputDir: 'storybook-static',
       mode: 'static',
       tsConfig: 'path/to/tsConfig.json',
+      webpackStatsJson: false,
+    });
+  });
+
+  it('should build storybook with webpack stats.json', async () => {
+    const run = await architect.scheduleBuilder('@storybook/angular:build-storybook', {
+      tsConfig: 'path/to/tsConfig.json',
+      compodoc: false,
+      webpackStatsJson: true,
+    });
+
+    const output = await run.result;
+
+    await run.stop();
+
+    expect(output.success).toBeTruthy();
+    expect(cpSpawnMock.spawn).not.toHaveBeenCalledWith();
+    expect(buildStandaloneMock).toHaveBeenCalledWith({
+      angularBrowserTarget: null,
+      angularBuilderContext: expect.any(Object),
+      angularBuilderOptions: {},
+      configDir: '.storybook',
+      loglevel: undefined,
+      quiet: false,
+      outputDir: 'storybook-static',
+      mode: 'static',
+      tsConfig: 'path/to/tsConfig.json',
+      webpackStatsJson: true,
     });
   });
 
   it('should throw error', async () => {
-    buildStandaloneMock.mockRejectedValue(new Error());
+    buildStandaloneMock.mockRejectedValue(true);
 
     const run = await architect.scheduleBuilder('@storybook/angular:start-storybook', {
       browserTarget: 'angular-cli:build-2',
@@ -162,6 +191,7 @@ describe('Build Storybook Builder', () => {
       outputDir: 'storybook-static',
       mode: 'static',
       tsConfig: './storybook/tsconfig.ts',
+      webpackStatsJson: false,
     });
   });
 
@@ -188,6 +218,7 @@ describe('Build Storybook Builder', () => {
       outputDir: 'storybook-static',
       mode: 'static',
       tsConfig: 'path/to/tsConfig.json',
+      webpackStatsJson: false,
     });
   });
 });