+++ docs/toc.js
@@ -180,6 +180,11 @@ module.exports = {
           type: 'link',
         },
         {
+          pathSegment: 'test-runner',
+          title: 'Test runner',
+          type: 'link',
+        },
+        {
           pathSegment: 'visual-testing',
           title: 'Visual tests',
           type: 'link',
@@ -219,6 +224,11 @@ module.exports = {
         {
           pathSegment: 'embed',
           title: 'Embed',
+          type: 'link',
+        },
+        {
+          pathSegment: 'design-integrations',
+          title: 'Design integrations',
           type: 'link',
         },
         {
@@ -348,15 +358,15 @@ module.exports = {
           type: 'link',
         },
         {
+          pathSegment: 'telemetry',
+          title: 'Telemetry',
+          type: 'link',
+        },
+        {
           pathSegment: '',
           title: 'Integration',
           type: 'menu',
           children: [
-            {
-              pathSegment: 'webpack',
-              title: 'Webpack',
-              type: 'link',
-            },
             {
               pathSegment: 'babel',
               title: 'Babel',
@@ -419,6 +429,33 @@ module.exports = {
         {
           pathSegment: 'environment-variables',
           title: 'Environment variables',
+          type: 'link',
+        },
+      ],
+    },
+    {
+      title: '🧰 Builders',
+      pathSegment: 'builders',
+      type: 'menu',
+      children: [
+        {
+          pathSegment: 'overview',
+          title: 'Introduction',
+          type: 'link',
+        },
+        {
+          pathSegment: 'vite',
+          title: 'Vite',
+          type: 'link',
+        },
+        {
+          pathSegment: 'webpack',
+          title: 'Webpack',
+          type: 'link',
+        },
+        {
+          pathSegment: 'builder-api',
+          title: 'API',
           type: 'link',
         },
       ],