+++ examples/angular-cli/src/stories/basics/component-with-ng-on-destroy/component-with-on-destroy.stories.ts
@@ -4,7 +4,7 @@ import { Story, Meta } from '@storybook/angular';
 @Component({
   selector: 'on-destroy',
   template: `Current time: {{ time }} <br />
-    📝 The current time in console should no longer display after a change of stroy`,
+    📝 The current time in console should no longer display after a change of story`,
 })
 class OnDestroyComponent implements OnInit, OnDestroy {
   time: string;