+++ examples/react-ts-webpack4/main.ts
@@ -19,6 +19,7 @@ const config: StorybookConfig = {
   ],
   core: {
     builder: 'webpack4',
+    disableTelemetry: true,
   },
   typescript: {
     check: true,