+++ examples/vue-3-cli/.storybook/main.js
@@ -7,8 +7,10 @@ module.exports = {
   ],
   core: {
     builder: 'webpack4',
+    disableTelemetry: true,
   },
   features: {
     buildStoriesJson: true,
+    channelOptions: { allowFunction: false, maxDepth: 10 },
   },
 };