+++ examples/vue-cli/.storybook/main.js
@@ -9,6 +9,8 @@ module.exports = {
   ],
   core: {
     builder: 'webpack4',
+    channelOptions: { allowFunction: false, maxDepth: 10 },
+    disableTelemetry: true,
   },
   features: {
     buildStoriesJson: true,