+++ lib/cli/src/frameworks/vue/Header.vue
@@ -21,6 +21,7 @@
         <h1>Acme</h1>
       </div>
       <div>
+        <span class="welcome" v-if="user">Welcome, <b>{{ user.name }}</b>!</span>
         <my-button size="small" @onClick="onLogout" label="Log out" v-if="user" />
         <my-button size="small" @onClick="onLogin" label="Log in" v-if="!user" />
         <my-button primary size="small" @onClick="onCreateAccount" label="Sign up" v-if="!user" />