+++ lib/cli/src/generators/METEOR/index.ts
@@ -3,7 +3,7 @@ import JSON5 from 'json5';
 import { baseGenerator, Generator } from '../baseGenerator';
 
 const generator: Generator = async (packageManager, npmOptions, options) => {
-  baseGenerator(packageManager, npmOptions, options, 'react', {
+  await baseGenerator(packageManager, npmOptions, options, 'react', {
     extraPackages: ['react', 'react-dom', '@babel/preset-env', '@babel/preset-react'],
     staticDir: 'dist',
   });