+++ lib/ui/src/app.stories.tsx
@@ -39,10 +39,11 @@ export const Default = () => (
       layout={{
         initialActive: 'addons',
         isFullscreen: false,
-        isToolshown: true,
+        showToolbar: true,
         panelPosition: 'right',
         showNav: true,
         showPanel: true,
+        showTabs: true,
       }}
       panelCount={0}
       docsOnly={false}
@@ -66,10 +67,11 @@ export const LoadingState = () => (
       layout={{
         initialActive: 'addons',
         isFullscreen: false,
-        isToolshown: true,
+        showToolbar: true,
         panelPosition: 'right',
         showNav: true,
         showPanel: true,
+        showTabs: true,
       }}
       panelCount={0}
       docsOnly={false}